fix figma
