+++ components/common/CharmEditor/components/common/EmbeddedInputPopup.tsx
@@ -12,6 +12,7 @@ interface InputProps {
 
 export function EmbeddedInputPopup(props: InputProps) {
   const autoOpen = props.node.marks.some((mark) => mark.type.name === 'tooltip-marker');
+
   return (
     <PopperPopup autoOpen={autoOpen} popupContent={<Box width={750}>{props.children}</Box>}>
       <ListItem
